requirements.txt
streamlit==1.18.1
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain==0.0.184
openai==0.27.6
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
pandas
Pillow
pdf2image
pytesseract
# Uncomment to use Hugging Face language models
# huggingface-hub==0.14.1

# Uncomment to use Instructor Embeddings
# InstructorEmbedding==1.0.1
# sentence-transformers==2.2.2
